requirements.txt
fastapi
uvicorn
google-genai
chromadb
pydantic
pydantic-settings
python-dotenv
numpy<2.0.0
